Traffic-Simulation-main/requirements.txt
flask
ultralytics
Pillow
matplotlib
